Conflicts:
	IGPE_Project/core/src/com/mygdx/game/model/entities/Entity.java
	IGPE_Project/core/src/com/mygdx/game/view/GameView.java

+++ IGPE_Project/core/src/com/mygdx/game/model/entities/Entity.java
@@ -83,7 +83,7 @@ public abstract class Entity implements Animated, Steerable<Vector2>, Collidable
 		return body.getPosition();
 	}
 	
-	public int getHealth() {
+	public float getHealth() {
 		return health;
 	}
